Resolve circular import
Fix missing file
