+++ kpops/components/base_components/kafka_app.py
@@ -12,9 +12,9 @@ from typing_extensions import override
 from kpops.component_handlers import get_handlers
 from kpops.components.base_components.cleaner import Cleaner
 from kpops.components.base_components.helm_app import HelmAppValues
-from kpops.components.base_components.models.topic import KafkaTopic, KafkaTopicStr
 from kpops.components.base_components.pipeline_component import PipelineComponent
 from kpops.components.common.streams_bootstrap import StreamsBootstrap
+from kpops.components.common.topic import KafkaTopic, KafkaTopicStr
 from kpops.config import get_config
 from kpops.utils.docstring import describe_attr
 from kpops.utils.pydantic import (
