+++ kpops/components/streams_bootstrap/__init__.py
@@ -1,6 +1,7 @@
 from kpops.components.common.streams_bootstrap import StreamsBootstrap
-from kpops.components.streams_bootstrap.producer.producer_app import ProducerApp
-from kpops.components.streams_bootstrap.streams.streams_app import StreamsApp
+
+from .producer.producer_app import ProducerApp
+from .streams.streams_app import StreamsApp
 
 __all__ = (
     "StreamsBootstrap",
