+++ kpops/components/streams_bootstrap/producer/producer_app.py
@@ -7,12 +7,12 @@ from kpops.components.base_components.kafka_app import (
     KafkaApp,
     KafkaAppCleaner,
 )
-from kpops.components.base_components.models.topic import (
+from kpops.components.common.streams_bootstrap import StreamsBootstrap
+from kpops.components.common.topic import (
     KafkaTopic,
     OutputTopicTypes,
     TopicConfig,
 )
-from kpops.components.common.streams_bootstrap import StreamsBootstrap
 from kpops.components.streams_bootstrap.app_type import AppType
 from kpops.components.streams_bootstrap.producer.model import ProducerAppValues
 from kpops.utils.docstring import describe_attr
