+++ kpops/components/streams_bootstrap/streams/streams_app.py
@@ -7,8 +7,8 @@ from typing_extensions import override
 from kpops.component_handlers.kubernetes.pvc_handler import PVCHandler
 from kpops.components.base_components.helm_app import HelmApp
 from kpops.components.base_components.kafka_app import KafkaApp, KafkaAppCleaner
-from kpops.components.base_components.models.topic import KafkaTopic
 from kpops.components.common.streams_bootstrap import StreamsBootstrap
+from kpops.components.common.topic import KafkaTopic
 from kpops.components.streams_bootstrap.app_type import AppType
 from kpops.components.streams_bootstrap.streams.model import (
     StreamsAppValues,
