+++ tests/components/test_producer_app.py
@@ -7,7 +7,7 @@ from pytest_mock import MockerFixture
 from kpops.component_handlers import ComponentHandlers, get_handlers
 from kpops.component_handlers.helm_wrapper.model import HelmUpgradeInstallFlags
 from kpops.component_handlers.helm_wrapper.utils import create_helm_release_name
-from kpops.components.base_components.models.topic import (
+from kpops.components.common.topic import (
     KafkaTopic,
     OutputTopicTypes,
     TopicConfig,
