+++ tests/pipeline/test_components/components.py
@@ -10,8 +10,8 @@ from kpops.components.base_components.models import ModelName, ModelVersion, Top
 from kpops.components.base_components.models.to_section import (
     ToSection,
 )
-from kpops.components.base_components.models.topic import OutputTopicTypes, TopicConfig
 from kpops.components.base_components.pipeline_component import PipelineComponent
+from kpops.components.common.topic import OutputTopicTypes, TopicConfig
 from kpops.components.streams_bootstrap.producer.producer_app import ProducerApp
 from kpops.components.streams_bootstrap.streams.streams_app import StreamsApp
