+++ tests/pipeline/test_components_without_schema_handler/components.py
@@ -2,8 +2,8 @@ from typing_extensions import override
 
 from kpops.component_handlers.kafka_connect.model import KafkaConnectorConfig
 from kpops.components.base_components.kafka_connector import KafkaSinkConnector
-from kpops.components.base_components.models.topic import OutputTopicTypes
 from kpops.components.base_components.pipeline_component import PipelineComponent
+from kpops.components.common.topic import OutputTopicTypes
 from kpops.components.streams_bootstrap.producer.producer_app import ProducerApp
 from kpops.components.streams_bootstrap.streams.streams_app import StreamsApp
